// client/typescript/models/FileSystem.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FileSystem = {
    readonly id: string;
    path: string;
    readonly depth: number | null;
    type?: string;
    ref?: string | null;
    href?: string | null;
    meta?: any;
    shortcut?: boolean | null;
    readonly created_at: string;
    readonly last_viewed_at: string | null;
};